+++ packages/shadcn/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @shadcn/ui
 
+## 4.0.5
+
+### Patch Changes
+
+- [#9960](https://github.com/shadcn-ui/ui/pull/9960) [`5ee456735377158c12cf55eefbe872f7303e1325`](https://github.com/shadcn-ui/ui/commit/5ee456735377158c12cf55eefbe872f7303e1325) Thanks [@shadcn](https://github.com/shadcn)! - update handling of init urls
+
+## 4.0.4
+
+### Patch Changes
+
+- [#9957](https://github.com/shadcn-ui/ui/pull/9957) [`aa841e35cf405e574123478b46f4058cff0e179a`](https://github.com/shadcn-ui/ui/commit/aa841e35cf405e574123478b46f4058cff0e179a) Thanks [@shadcn](https://github.com/shadcn)! - fix cache in resolveRegistryBaseConfig
+
+## 4.0.3
+
+### Patch Changes
+
+- [#9950](https://github.com/shadcn-ui/ui/pull/9950) [`ce2c3ca688916db4fdb5fe173e9c6902f78696ed`](https://github.com/shadcn-ui/ui/commit/ce2c3ca688916db4fdb5fe173e9c6902f78696ed) Thanks [@shadcn](https://github.com/shadcn)! - add support for translucent menu
+
 ## 4.0.2
 
 ### Patch Changes